Made even more similar to json.Unmarshal output.
Instead of building a tree out of FormTree and Slice instances, I've built
the tree from map[string]interface{} and []interface{} so that it can be
processed by anything that might process the output of json.Unmarshal to
an empty interface.

The (renamed) accessors now return FormTree and Slice instances so that
FormTree still has the kind of interface I originally envisioned while
the underlying data structure is amenable to being used in place of a
json.Unmarshal result.